server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["DMITRII <dengivseti@list.ru>"]
packages = [
    { include="scripts", from="." },
]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.19.1"
loguru = "^0.5.3"
fastapi = "^0.70.0"
ormar = {extras = ["postgresql"], version = "^0.10.20"}
odmantic = "^0.3.5"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
fastapi-users = {extras = ["ormar"], version = "^8.1.1"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
serve = 'scripts.run_server:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

